refactor: rename header files (path, log, define, calibration)
