+++ retinify/include/retinify/paths.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "retinify/define.hpp"
+#include "retinify/attributes.hpp"
 
 namespace retinify
 {
